rootdir: update for our device

+++ rootdir/bin/init.qcom.usb.sh
@@ -43,21 +43,15 @@ esoc_name=`cat /sys/bus/esoc/devices/esoc0/esoc_name 2> /dev/null`
 
 target=`getprop ro.board.platform`
 
-#
-# Allow USB enumeration with default PID/VID
-#
-baseband=`getprop ro.baseband`
-
-echo 1  > /sys/class/android_usb/f_mass_storage/lun/nofua
-usb_config=`getprop persist.vendor.usb.config`
-debuggable=`getprop ro.debuggable`
-buildvariant=`getprop ro.build.type`
-factorybuild=`getprop ro.boot.factorybuild`
+if [ -f /sys/class/android_usb/f_mass_storage/lun/nofua ]; then
+	echo 1  > /sys/class/android_usb/f_mass_storage/lun/nofua
+fi
 
 #
 # Override USB default composition
 #
 # If USB persist config not set, set default configuration
+build_type=`getprop ro.build.type`
 if [ "$(getprop persist.vendor.usb.config)" == "" -a \
 	"$(getprop init.svc.vendor.usb-gadget-hal-1-0)" != "running" ]; then
     if [ "$esoc_name" != "" ]; then
@@ -72,33 +66,6 @@ if [ "$(getprop persist.vendor.usb.config)" == "" -a \
 	          "Dragon" | "SBC")
 	              setprop persist.vendor.usb.config diag,adb
 	          ;;
-		  "DAVINCI")
-                      if [ "$buildvariant" = "eng" ]; then
-                         setprop persist.vendor.usb.config diag,serial_cdev,rmnet,dpl,qdss,adb
-                      elif [ -z "$debuggable" -o "$debuggable" = "1"  ]; then
-                         setprop persist.vendor.usb.config adb
-                      else
-                         setprop persist.vendor.usb.config none
-                      fi
-                  ;;
-                  "TUCANA" | "TOCO")
-                      if [ "$buildvariant" = "eng" -o "$factorybuild" = "1" ]; then
-                         setprop persist.vendor.usb.config diag,serial_cdev,rmnet,dpl,qdss,adb
-                      elif [ -z "$debuggable" -o "$debuggable" = "1"  ]; then
-                         setprop persist.vendor.usb.config adb
-                      else
-                         setprop persist.vendor.usb.config none
-                      fi
-                  ;;
-		  "PHOENIX")
-                      if [ "$buildvariant" = "eng" -o "$buildvariant" = "userdebug" ]; then
-                         setprop persist.vendor.usb.config diag,serial_cdev,rmnet,dpl,qdss,adb
-                      elif [ -z "$debuggable" -o "$debuggable" = "1"  ]; then
-                         setprop persist.vendor.usb.config adb
-                      else
-                         setprop persist.vendor.usb.config none
-                      fi
-                  ;;
                   *)
 		  case "$soc_machine" in
 		    "SA")
@@ -106,9 +73,6 @@ if [ "$(getprop persist.vendor.usb.config)" == "" -a \
 		    ;;
 		    *)
 	            case "$target" in
-	              "msm8996")
-	                  setprop persist.vendor.usb.config diag,serial_cdev,serial_tty,rmnet_ipa,mass_storage,adb
-		      ;;
 	              "msm8909")
 		          setprop persist.vendor.usb.config diag,serial_smd,rmnet_qti_bam,adb
 		      ;;
@@ -133,14 +97,23 @@ if [ "$(getprop persist.vendor.usb.config)" == "" -a \
 				      setprop persist.vendor.usb.config diag,serial_smd,rmnet_ipa,adb
 			      fi
 		      ;;
+	              "msm8996")
+			      if [ -d /config/usb_gadget ]; then
+				      setprop persist.vendor.usb.config diag,serial_cdev,rmnet,adb
+			      else
+				      setprop persist.vendor.usb.config diag,serial_cdev,serial_tty,rmnet_ipa,mass_storage,adb
+			      fi
+		      ;;
 	              "msm8998" | "sdm660" | "apq8098_latv")
 		          setprop persist.vendor.usb.config diag,serial_cdev,rmnet,adb
 		      ;;
 	              "sdm845" | "sdm710")
 		          setprop persist.vendor.usb.config diag,serial_cdev,rmnet,dpl,adb
 		      ;;
-	              "msmnile" | "sm6150" | "trinket" | "lito")
+	              "msmnile" | "sm6150" | "trinket")
+			if [ "$build_type" != "user" ]; then
 			  setprop persist.vendor.usb.config diag,serial_cdev,rmnet,dpl,qdss,adb
+			fi
 		      ;;
 	              *)
 		          setprop persist.vendor.usb.config diag,adb
@@ -179,30 +152,44 @@ if [ "$target" == "msm8937" ]; then
 		*)
 		;;
 	   esac
+	else
+	   case "$soc_id" in
+		"313" | "320")
+		   setprop vendor.usb.rndis.func.name "rndis_bam"
+		   setprop vendor.usb.rmnet.func.name "rmnet_bam"
+		   setprop vendor.usb.rmnet.inst.name "rmnet"
+		   setprop vendor.usb.dpl.inst.name "dpl"
+		;;
+		*)
+		;;
+	   esac
 	fi
+fi
+
+if [ "$target" == "msm8996" ]; then
+       if [ -d /config/usb_gadget ]; then
+                  setprop vendor.usb.rndis.func.name "rndis_bam"
+                  setprop vendor.usb.rmnet.func.name "rmnet_bam"
+                  setprop vendor.usb.rmnet.inst.name "rmnet"
+                  setprop vendor.usb.dpl.inst.name "dpl"
+       fi
 fi
 
 # check configfs is mounted or not
 if [ -d /config/usb_gadget ]; then
+	# Chip-serial is used for unique MSM identification in Product string
+	msm_serial=`cat /sys/devices/soc0/serial_number`;
+	msm_serial_hex=`printf %08X $msm_serial`
+	machine_type=`cat /sys/devices/soc0/machine`
+	product_string="$machine_type-$soc_hwplatform _SN:$msm_serial_hex"
+	echo "$product_string" > /config/usb_gadget/g1/strings/0x409/product
+
 	# ADB requires valid iSerialNumber; if ro.serialno is missing, use dummy
 	serialnumber=`cat /config/usb_gadget/g1/strings/0x409/serialnumber 2> /dev/null`
 	if [ "$serialnumber" == "" ]; then
 		serialno=1234567
 		echo $serialno > /config/usb_gadget/g1/strings/0x409/serialnumber
 	fi
-
-	persist_comp=`getprop persist.vendor.usb.config`
-	comp=`getprop sys.usb.config`
-	echo $persist_comp
-	echo $comp
-	if [ "$comp" != "$persist_comp" ]; then
-		echo "setting sys.usb.config"
-		setprop sys.usb.config $persist_comp
-	fi
-
-	setprop sys.usb.configfs 1
-
-	setprop vendor.usb.configfs 1
 fi
 
 #
